+++ src/main/features/response/form/models/howMuchOwed.ts
@@ -1,7 +1,7 @@
 import { Serializable } from 'models/serializable'
 import { IsDefined, IsPositive, MaxLength } from 'class-validator'
 import { IsNotBlank } from 'app/forms/validation/validators/isBlank'
-import { ValidationErrors } from 'features/validationErrors'
+import { ValidationErrors } from 'forms/validation/validationErrors'
 import { Fractions } from 'forms/validation/validators/fractions'
 import { ValidationConstraints } from 'forms/validation/validationConstraints'
