+++ src/main/features/response/routes/impact-of-dispute.ts
@@ -5,6 +5,7 @@ import { ErrorHandling } from 'common/errorHandling'
 import { Form } from 'forms/form'
 import { ImpactOfDispute } from 'response/form/models/impactOfDispute'
 
+/* tslint:disable:no-default-export */
 export default express.Router()
   .get(Paths.impactOfDisputePage.uri,
     ErrorHandling.apply(async (req: express.Request, res: express.Response, next: express.NextFunction) => {
