+++ src/test/features/response/form/models/howMuchOwed.ts
@@ -6,7 +6,7 @@ import * as randomstring from 'randomstring'
 import { Validator } from 'class-validator'
 import { expectValidationError } from '../../../../app/forms/models/validationUtils'
 import { HowMuchOwed } from 'features/response/form/models/howMuchOwed'
-import { ValidationErrors } from 'features/validationErrors'
+import { ValidationErrors } from 'forms/validation/validationErrors'
 import { ValidationConstraints } from 'forms/validation/validationConstraints'
 
 describe('HowMuchOwed', () => {
